+++ src/user/syscalls.c
@@ -1,11 +1,15 @@
+#include <stddef.h>
+
 #pragma GCC diagnostic ignored "-Wunused-parameter"
 #pragma GCC diagnostic ignored "-Wreturn-type"
 
-int Reply(int tid, const char* reply, int rplen) { __asm__ volatile("swi #7"); }
+int Reply(int tid, const char* reply, size_t rplen) {
+    __asm__ volatile("swi #7");
+}
 
-int Receive(int* tid, char* msg, int msglen) { __asm__ volatile("swi #6"); }
+int Receive(int* tid, char* msg, size_t msglen) { __asm__ volatile("swi #6"); }
 
-int Send(int tid, const char* msg, int msglen, char* reply, int rplen) {
+int Send(int tid, const char* msg, size_t msglen, char* reply, size_t rplen) {
     __asm__ volatile("swi #5");
 }
